requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
SQLAlchemy>=2.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
WTForms>=3.0.0
Werkzeug>=2.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
